# openapi.yaml
openapi: 3.1.0
info:
  title: Collider Custody - Transaction Security Layer
  description: |
    On-prem Transaction Security Layer + Wallet-as-a-Service (Ethereum)
    
    ## Features
    - Wallet Registry (WaaS) - Create and manage Ethereum wallets
    - Transaction Orchestrator - State machine for transaction lifecycle
    - KYT (Know Your Transaction) - Screen transactions
    - Policy Engine - Enforce limits and rules
    - Approvals with SoD - Multi-approval workflows
    - Signing Adapter - Dev mode signer
    - Chain Listener - Monitor confirmations
    - Audit Log - Tamper-evident hash-chain
  version: 1.0.0
  contact:
    name: Collider Team
servers:
  - url: http://localhost:8000
    description: Local development
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    CorrelationId:
      name: X-Correlation-ID
      in: header
      required: false
      schema:
        type: string
      description: Request correlation ID for tracing
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
      description: Idempotency key for POST requests
  schemas:
    WalletType:
      type: string
      enum: [RETAIL, TREASURY, OPS, SETTLEMENT]
    RiskProfile:
      type: string
      enum: [LOW, MEDIUM, HIGH]
    WalletRoleType:
      type: string
      enum: [OWNER, OPERATOR, VIEWER, APPROVER]
...
    WalletCreate:
      type: object
      required:
        - wallet_type
        - subject_id
      properties:
        wallet_type:
          $ref: '#/components/schemas/WalletType'
        subject_id:
          type: string
          minLength: 1
          maxLength: 255
        tags:
          type: object
        risk_profile:
          $ref: '#/components/schemas/RiskProfile'
          default: MEDIUM
...
    WalletRoleAssign:
      type: object
      required:
        - user_id
        - role
      properties:
        user_id:
          type: string
          format: uuid
        role:
          $ref: '#/components/schemas/WalletRoleType'
...
    CorrelatedResponse:
      type: object
      properties:
        correlation_id:
          type: string
        data:
          type: object
        timestamp:
          type: string
          format: date-time
...
paths:
...
  /v1/wallets:
    post:
      tags: [Wallets]
      summary: Create wallet
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WalletCreate'
      responses:
        '200':
          description: Wallet created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CorrelatedResponse'
    get:
      tags: [Wallets]
      summary: List wallets
      parameters:
        - name: wallet_type
          in: query
          schema:
            $ref: '#/components/schemas/WalletType'
        - name: subject_id
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of wallets
  /v1/wallets/{wallet_id}/roles:
    post:
      tags: [Wallets]
      summary: Assign role to wallet
      parameters:
        - name: wallet_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WalletRoleAssign'
      responses:
        '200':
          description: Role assigned
...